+++ supreme/src/commonTest/kotlin/at/asitplus/signum/provider/ParallelismTest.kt
@@ -1,11 +1,12 @@
 package at.asitplus.signum.provider
 
 import at.asitplus.signum.supreme.os.SigningProviderI
-import at.asitplus.testballoon.matrix.matrixSuite
+import at.asitplus.testballoon.invoke
+import de.infix.testBalloon.framework.core.testSuite
 import kotlinx.coroutines.joinAll
 import kotlinx.coroutines.launch
 
-val ParallelismTest by matrixSuite {
+val ParallelismTest by testSuite {
     val alias = "TestAlias"
 
     "10k Loop" {